refactor sync relevant python tests
