+++ tests/sync_auto_test.py
@@ -3,17 +3,15 @@
 from test_framework.test_framework import TestFramework
 from utility.utils import wait_until
 
-class AutoRandomSyncTest(TestFramework):
+class AutoSyncTest(TestFramework):
     def setup_params(self):
         self.num_nodes = 2
 
-        # Enable random auto sync only
+        # Enable auto sync
         for i in range(self.num_nodes):
             self.zgs_node_configs[i] = {
                 "sync": {
-                    "auto_sync_enabled": True,
-                    "max_sequential_workers": 0,
-                    "max_random_workers": 3,
+                    "auto_sync_enabled": True
                 }
             }
 
@@ -29,4 +27,4 @@ class AutoRandomSyncTest(TestFramework):
         wait_until(lambda: self.nodes[1].zgs_get_file_info(data_root_2)["finalized"])
 
 if __name__ == "__main__":
-    AutoRandomSyncTest().main()
+    AutoSyncTest().main()
